backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Azure OpenAI (only for chat, not embeddings)
openai==1.3.7
azure-identity==1.15.0

# Vector Database and Embeddings - Following notebook approach
chromadb==0.4.18
sentence-transformers==2.2.2

# Data Processing - NumPy 1.x for compatibility
pandas>=1.5.0,<2.2.0
numpy>=1.24.0,<2.0.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1

# HTTP Client and API
aiohttp==3.9.1
httpx==0.25.2
requests==2.31.0

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# YAML and JSON Processing
pyyaml==6.0.1
jsonschema==4.20.0

# Logging and Monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Security and Validation
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
textstat==0.7.3

# Machine Learning for embeddings
torch>=1.9.0
transformers>=4.6.0
tokenizers>=0.10.0